More restructuring and renaming
